refactor: fix module import pattern
